@@ -3,10 +3,9 @@
 namespace Honeybadger\HoneybadgerLaravel\Commands;
 
 use Exception;
-use Honeybadger\Honeybadger;
-use Illuminate\Console\Command;
 use Honeybadger\Contracts\Reporter;
 use Honeybadger\HoneybadgerLaravel\Exceptions\TestException;
+use Illuminate\Console\Command;
 
 class HoneybadgerTestCommand extends Command
 {
@@ -24,21 +23,24 @@ class HoneybadgerTestCommand extends Command
      */
     protected $description = 'Tests notifications to Honeybadger';
 
-    /**
-     * Execute the console command.
-     *
-     * @return mixed
-     */
-    public function handle(Reporter $honeybadger)
+    public function handle()
     {
+        /** @var Reporter $honeybadger */
+        $honeybadger = app('honeybadger.loud');
+
         try {
+            if (! config('honeybadger.report_data')) {
+                $this->info("You have `report_data` set to false in your config. Errors won't be reported in this environment.");
+                $this->info("We've switched it to true for this test, but you should check that it's enabled for your production environments.");
+            }
             $result = $honeybadger->notify(new TestException);
-            $this->info('A test exception was sent to Honeybadger');
-            if (is_null(array_get($result, 'id'))) {
+            $id = $result['id'] ?? null;
+            if (is_null($id)) {
                 throw new Exception('There was an error sending the exception to Honeybadger');
             }
 
-            $this->line(sprintf('https://app.honeybadger.io/notice/', array_get($result, 'id')));
+            $noticeUrl = "https://app.honeybadger.io/notice/$id";
+            $this->info("Successfully sent a test exception to Honeybadger: $noticeUrl");
         } catch (Exception $e) {
             $this->error($e->getMessage());
         }
