@@ -11,33 +11,33 @@ interface Installer
      * @param  string  $filePath
      * @return bool
      */
-    public function writeConfig(array $config, string $filePath) : bool;
+    public function writeConfig(array $config, string $filePath): bool;
 
     /**
      * Publish the config file for Laravel.
      *
      * @return bool
      */
-    public function publishLaravelConfig() : bool;
+    public function publishLaravelConfig(): bool;
 
     /**
      * Publish the config file for Lumen.
      *
      * @return bool
      */
-    public function publishLumenConfig() : bool;
+    public function publishLumenConfig(): bool;
 
     /**
      * Whether the config file needs to be published or not.
      *
      * @return bool
      */
-    public function shouldPublishConfig() : bool;
+    public function shouldPublishConfig(): bool;
 
     /**
      * Send a test exception to Honeybadger.
      *
      * @return array
      */
-    public function sendTestException() : array;
+    public function sendTestException(): array;
 }
